feat(nuxt): support universal global middleware
