+++ examples/routing/universal-router/middleware/always-run.global.ts
@@ -1,0 +1,3 @@
+export default defineNuxtRouteMiddleware(() => {
+  console.log('running global middleware')
+})
