+++ packages/nuxt/src/core/templates.ts
@@ -5,6 +5,7 @@ import { genArrayFromRaw, genDynamicImport, genExport, genImport, genObjectFromR
 import { isAbsolute, join, relative } from 'pathe'
 import { resolveSchema, generateTypes } from 'untyped'
 import escapeRE from 'escape-string-regexp'
+import { getImportName } from './utils'
 
 export interface TemplateContext {
   nuxt: Nuxt
@@ -159,13 +160,29 @@ export const schemaTemplate = {
 // Add layouts template
 export const layoutTemplate: NuxtTemplate = {
   filename: 'layouts.mjs',
-  getContents ({ app }) {
+  getContents ({ app }: TemplateContext) {
     const layoutsObject = genObjectFromRawEntries(Object.values(app.layouts).map(({ name, file }) => {
       return [name, `defineAsyncComponent(${genDynamicImport(file)})`]
     }))
     return [
       'import { defineAsyncComponent } from \'vue\'',
           `export default ${layoutsObject}`
+    ].join('\n')
+  }
+}
+
+// Add middleware template
+export const middlewareTemplate: NuxtTemplate = {
+  filename: 'middleware.mjs',
+  getContents ({ app }: TemplateContext) {
+    const middleware = Object.values(app.middleware)
+    const globalMiddleware = middleware.filter(mw => mw.global)
+    const namedMiddleware = middleware.filter(mw => !mw.global)
+    const namedMiddlewareObject = genObjectFromRawEntries(namedMiddleware.map(mw => [mw.name, genDynamicImport(mw.path)]))
+    return [
+      ...globalMiddleware.map(mw => genImport(mw.path, getImportName(mw.name))),
+          `export const globalMiddleware = ${genArrayFromRaw(globalMiddleware.map(mw => getImportName(mw.name)))}`,
+          `export const namedMiddleware = ${namedMiddlewareObject}`
     ].join('\n')
   }
 }
